add page and controller .

+++ swallow-web/src/main/java/com/dianping/swallow/web/model/alarm/QPSAlarmSetting.java
@@ -29,7 +29,7 @@ public class QPSAlarmSetting {
 		this.valley = valley;
 	}
 
-	public long getFluctuation() {
+	public int getFluctuation() {
 		return fluctuation;
 	}
